feat: Implement admin UI for category and user management, migrate various entity IDs to UUIDs or Longs, and refine DAO/service layer interactions.

+++ src/main/java/com/polycoffee/dao/impl/CategoryDAO.java
@@ -23,7 +23,7 @@ public class CategoryDAO implements ICategoryDAO {
     }
 
     @Override
-    public Categories findById(Integer id) {
+    public Categories findById(Long id) {
         EntityManager em = XJPA.createEntityManager();
         try {
             return em.find(Categories.class, id);
@@ -63,7 +63,7 @@ public class CategoryDAO implements ICategoryDAO {
     }
 
     @Override
-    public void delete(Integer id) {
+    public void delete(Long id) {
         EntityManager em = XJPA.createEntityManager();
         try {
             em.getTransaction().begin();
